requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
lpips>=0.1.4
tqdm>=4.64.0
pyyaml>=6.0
scipy>=1.7.0
